Rename `ContentForCompletionProvider`
