+++ packages/theme-language-server-common/src/completions/providers/index.ts
@@ -1,4 +1,4 @@
-export { ContentForBlockCompletionProvider } from './ContentForBlockCompletionProvider';
+export { ContentForCompletionProvider } from './ContentForCompletionProvider';
 export { HtmlTagCompletionProvider } from './HtmlTagCompletionProvider';
 export { HtmlAttributeCompletionProvider } from './HtmlAttributeCompletionProvider';
 export { HtmlAttributeValueCompletionProvider } from './HtmlAttributeValueCompletionProvider';
